Refactor: Flatten module hierarchy for simplicity
- Flatten `src/grammar/` to `src/grammar.rs` and `src/grammar.pest`.
- Flatten `src/ast/` to `src/ast.rs`.
- Flatten `src/errors/` to `src/errors.rs`.
- Flatten `src/parser/` to `src/parser.rs`.
- Move `src/experimental/bard.rs` to `src/highlight.rs`.
- Update `tests/lambda_tests.rs` to expect standard Rust method names (e.g., `.map` instead of `.g_map`) for iterator operations, aligning tests with the actual code generation logic.
- Remove empty directories.

This reduction simplifies the project structure, reduces file count, and eliminates unnecessary nesting, adhering to the Razor philosophy.
