+++ src/highlight.rs
@@ -19,7 +19,7 @@
 //! # Usage
 //!
 //! ```rust
-//! use glossa::experimental::bard::highlight;
+//! use glossa::highlight::highlight;
 //!
 //! let source = "ὁ ἄνθρωπος τὸν λόγον λέγει.";
 //! let highlighted = highlight(source).unwrap();
