+++ src/lib.rs
@@ -46,8 +46,8 @@
 pub mod ast;
 pub mod codegen;
 pub mod errors;
-pub mod experimental;
 pub mod grammar;
+pub mod highlight;
 pub mod morphology;
 pub mod parser;
 pub mod semantic;
