+++ src/parser.rs
@@ -1,7 +1,19 @@
-//! AST Builder
+//! Semantic Parser for ΓΛΩΣΣΑ
 //!
-//! This module is responsible for converting the raw Concrete Syntax Tree (CST)
-//! produced by the Pest parser into a strongly-typed Abstract Syntax Tree (AST).
+//! This module bridges the gap between the raw text parsing (Grammar) and the
+//! high-level program structure (AST).
+//!
+//! # The Parsing Flow
+//!
+//! 1. **Grammar (`src/grammar`)**: Uses [`pest`] (PEG parser) to tokenize the input
+//!    and verify it matches the language rules. This produces a "Concrete Syntax Tree" (CST)
+//!    of untyped pairs (e.g., `Rule::greek_word`, `Rule::number_literal`).
+//!
+//! 2. **Builder (`src/parser.rs`)**: Walks this CST and constructs strongly-typed
+//!    [`crate::ast`] nodes. This is where we handle:
+//!    * Text normalization (converting `Ἀθῆναι` to `αθηναι`)
+//!    * Number parsing
+//!    * Structural validation (e.g., ensuring a trait method has a name)
 //!
 //! # Safety: Recursion Depth
 //!
@@ -11,15 +23,58 @@
 //! recursive descent parsing phase.
 
 use crate::ast::*;
-use crate::grammar::{Rule, parse};
+use crate::errors::GlossaError;
+use crate::grammar::{Rule, parse as grammar_parse};
 use pest::iterators::Pair;
 
-/// Build an AST from source code
-pub fn parse_source(source: &str) -> Result<Program, ParseError> {
+/// Parse a ΓΛΩΣΣΑ source string into an AST
+///
+/// This is the main entry point for the parsing phase.
+///
+/// # Examples
+///
+/// ```
+/// use glossa::parser::parse;
+///
+/// let source = "«χαῖρε» λέγε.";
+/// let program = parse(source).unwrap();
+/// assert_eq!(program.statements.len(), 1);
+/// ```
+pub fn parse(source: &str) -> Result<Program, GlossaError> {
+    parse_source(source).map_err(GlossaError::from)
+}
+
+impl From<ParseError> for GlossaError {
+    fn from(err: ParseError) -> Self {
+        GlossaError::parse(err.to_string())
+    }
+}
+
+/// Errors that can occur during AST construction
+#[derive(Debug, Clone, thiserror::Error)]
+pub enum ParseError {
+    #[error("Parse error: {0}")]
+    PestError(String),
+
+    #[error("Empty term in expression")]
+    EmptyTerm,
+
+    #[error("Invalid number: {0}")]
+    InvalidNumber(String),
+
+    #[error("Unexpected rule: {0}")]
+    UnexpectedRule(String),
+
+    #[error("Recursion limit exceeded: depth > {0}")]
+    RecursionLimitExceeded(usize),
+}
+
+/// Build an AST from source code (Internal)
+fn parse_source(source: &str) -> Result<Program, ParseError> {
     // Check recursion depth before parsing to prevent stack overflow
     check_recursion_depth(source)?;
 
-    let pairs = parse(source).map_err(|e| ParseError::PestError(e.to_string()))?;
+    let pairs = grammar_parse(source).map_err(|e| ParseError::PestError(e.to_string()))?;
 
     let mut statements = Vec::new();
 
@@ -584,25 +639,6 @@ fn build_array_element(pair: Pair<'_, Rule>) -> Result<Expr, ParseError> {
     }
 }
 
-/// Errors that can occur during AST construction
-#[derive(Debug, Clone, thiserror::Error)]
-pub enum ParseError {
-    #[error("Parse error: {0}")]
-    PestError(String),
-
-    #[error("Empty term in expression")]
-    EmptyTerm,
-
-    #[error("Invalid number: {0}")]
-    InvalidNumber(String),
-
-    #[error("Unexpected rule: {0}")]
-    UnexpectedRule(String),
-
-    #[error("Recursion limit exceeded: depth > {0}")]
-    RecursionLimitExceeded(usize),
-}
-
 /// Check recursion depth to prevent stack overflows
 ///
 /// This function performs a fast linear scan of the source code to ensure that
@@ -693,7 +729,7 @@ mod tests {
     #[test]
     fn test_parse_source_hello() {
         let source = "«χαῖρε» λέγε.";
-        let ast = parse_source(source).unwrap();
+        let ast = parse(source).unwrap();
 
         assert_eq!(ast.statements.len(), 1);
         assert!(!ast.statements[0].is_query());
@@ -702,7 +738,7 @@ mod tests {
     #[test]
     fn test_parse_source_string_literal() {
         let source = "«χαῖρε κόσμε» λέγε.";
-        let ast = parse_source(source).unwrap();
+        let ast = parse(source).unwrap();
 
         let expr = first_expr(&ast.statements[0]);
         if let Expr::Phrase(terms) = expr {
@@ -715,7 +751,7 @@ mod tests {
     #[test]
     fn test_parse_source_variable_binding() {
         let source = "ξ πέντε ἔστω.";
-        let ast = parse_source(source).unwrap();
+        let ast = parse(source).unwrap();
 
         assert_eq!(ast.statements.len(), 1);
         let expr = first_expr(&ast.statements[0]);
@@ -734,7 +770,7 @@ mod tests {
     #[test]
     fn test_parse_source_number_literal() {
         let source = "42 λέγε.";
-        let ast = parse_source(source).unwrap();
+        let ast = parse(source).unwrap();
 
         let expr = first_expr(&ast.statements[0]);
         if let Expr::Phrase(terms) = expr {
@@ -745,7 +781,7 @@ mod tests {
     #[test]
     fn test_parse_source_query() {
         let source = "ξ?";
-        let ast = parse_source(source).unwrap();
+        let ast = parse(source).unwrap();
 
         assert!(ast.statements[0].is_query());
     }
@@ -753,7 +789,7 @@ mod tests {
     #[test]
     fn test_parse_source_multiple_statements() {
         let source = "ξ πέντε ἔστω. ξ λέγε.";
-        let ast = parse_source(source).unwrap();
+        let ast = parse(source).unwrap();
 
         assert_eq!(ast.statements.len(), 2);
     }
@@ -761,7 +797,7 @@ mod tests {
     #[test]
     fn test_word_normalization() {
         let source = "χρήστου ὄνομα λέγε.";
-        let ast = parse_source(source).unwrap();
+        let ast = parse(source).unwrap();
 
         let expr = first_expr(&ast.statements[0]);
         if let Expr::Phrase(terms) = expr
@@ -776,7 +812,7 @@ mod tests {
     fn test_parse_source_with_comma() {
         // Test that commas create multiple clauses
         let source = "εἰ ξ μεῖζον, «ναί» λέγε.";
-        let ast = parse_source(source).unwrap();
+        let ast = parse(source).unwrap();
 
         assert_eq!(ast.statements.len(), 1);
         assert_eq!(ast.statements[0].clauses().len(), 2); // Two clauses separated by comma
